PCAngsd Result
To ensure the confidentiality of our results, here is a representative plot that I generated through the application of PCAngsd and R for conducting a principal component analysis. I opted to use color coordination based on estimated average depth to visually demonstrate that this metric does not influence the distribution of points.